Cadastro de Cliente quase completo, fazer teste de usuario já existente

+++ CAMADAS/MODEL/Cliente.cs
@@ -18,7 +18,7 @@ namespace Mercado_Thay_and_Lusca.CAMADAS.MODEL
 
         public string endereco { get; set; }
 
-        public float saldo { get; set; }
+        public double saldo { get; set; }
 
         public string imagem { get; set; }
     }
